FIX: Ensure user list updates live based on new data (#36)
For some reason, the modern ember @computed helper doesn't work in production. Changes to `online.users` were not triggering a re-evaluation of the component's properties. It does seem to work in development via ember CLI.

Some more investigation is required, but for now this commit reverts to the discourseComputed syntax