FIX: cannot read property of undefined (#119)
When upgrading this component to use the Glimmer topic list, we
converted @discourseComputed into value transformers but there's a bit
of magic with @discourseComputed when it comes to handling `.` and
`null` / `undefined` values.

This ensures the same behaviour and prevents the JS error from happening
when creating a new topic from the topic list.

Meta - https://meta.discourse.org/t/349416